third_party/WebKit/LayoutTests/fast/dom/shadow/olderShadowRoot-expected.txt

CONSOLE WARNING: line 12: Calling Element.createShadowRoot() for an element which already hosts a shadow root is deprecated and will be removed in M63, around December 2017. See https://www.chromestatus.com/features/4668884095336448 for more details.
Tests ShadowRoot.olderShadowRoot property

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".




PASS roots[1].olderShadowRoot is roots[0]
PASS roots[0].olderShadowRoot is null
PASS roots[2].olderShadowRoot is roots[1]
PASS roots[1].olderShadowRoot is roots[0]
PASS roots[0].olderShadowRoot is null
PASS roots[0].olderShadowRoot is null
PASS roots[1].olderShadowRoot is roots[0]
PASS roots[6].olderShadowRoot is roots[5]
PASS root.olderShadowRoot is null
PASS roots[1].olderShadowRoot is roots[0]



PASS successfullyParsed is true

TEST COMPLETE
